requirements.txt
pandas
tqdm
ipykernel
requests
python-dotenv